requirements.txt
requests
flask
google-generativeai